pyproject.toml
[project]
name = "animated-face-frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.2.1",
    "numpy>=2.3.4",
    "openai-whisper>=20250625",
    "pyaudio>=0.2.14",
    "pygame>=2.6.1",
    "pyyaml>=6.0.3",
    "torch>=2.9.0",
]
